Removed constructor & made DB truly read-only
The `RandomizingForLoadoutComboWhenAllWeaponsForThatComboAreDisabledThrows` showed that the DbFixture wasn't exactly read-only as expected. It was instantiated for each separate unit test method, but inside the unit-test variants with data from MemberData methods it turns out that the DbFixture wasn't instantiated for each unit-test variant. Simply instantiating a new DatabaseContext() instead of using a singleton fixed it, however doing so made the Generator's second constructor obsolete. Also changed the default connectionMode to readOnly.